+++ bestmobabot/__main__.py
@@ -1,6 +1,6 @@
 import os
 import signal
-from datetime import datetime, timedelta
+from datetime import datetime
 from typing import TextIO
 
 import click
@@ -24,6 +24,7 @@ from bestmobabot.vk import VK
 @click.option('is_trainer', '--trainer', help='Automatically train arena model once a day.', envvar='IS_TRAINER', is_flag=True)
 @click.option('raids', '--raid', help='Raid the mission specified by its ID and number of raids per day.', envvar='RAIDS', type=(str, int), multiple=True)
 @click.option('shops', '--shop', help='Buy goods specified by shop_id and slot_id every day', envvar='SHOPS', type=(str, str), multiple=True)
+@click.option('friend_ids', '--friend', help='Send daily gift to a friend specified by its ID.', envvar='FRIENDS', type=str, multiple=True)
 @click.option('--arena-early-stop', help='Minimum win probability to stop (grand) arena enemy search early.', envvar='ARENA_EARLY_STOP', type=float, default=0.95, show_default=True)
 @click.option('--arena-offset', help='Arena schedule offset in seconds.', envvar='ARENA_OFFSET', type=int, default=0, show_default=True)
 @click.option('--arena-teams-limit', help='Greater: better arena attackers but uses more resources.', envvar='ARENA_TEAMS_LIMIT', type=int, default=20000, show_default=True)
@@ -40,7 +41,7 @@ def main(remixsid: str, vk_token: str, log_file: TextIO, verbose: bool, arena_of
     get_translations()  # prefetch game translations
 
     with Database(constants.DATABASE_NAME) as db, API(db, remixsid) as api:
-        with Bot(db, api, VK(vk_token), arena_offset=timedelta(seconds=arena_offset), **kwargs) as bot:
+        with Bot(db, api, VK(vk_token), **kwargs) as bot:
             api.start()
             bot.start()
             logger.info(f'👋 Welcome {bot.user.name}! Your game time is {datetime.now(bot.user.tz):%H:%M:%S}.')