+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `ARENA_EARLY_STOP` parameter
 * Add `GRAND_ARENA_GENERATIONS` parameter
 * Add `ARENA_TEAMS_LIMIT` parameter
+* Add `FRIENDS` parameter
 * Model trainer is now using T-test to compare scores – it should decrease `n_estimators` and improve speed
 * Caching selected arena attackers during the search
 * Delete `test` branch